refactor: 공통으로 사용되는 dto 분리(#23)
