+++ src/main/java/play/pluv/oauth/spotify/dto/ThumbNailResponse.java
@@ -1,0 +1,9 @@
+package play.pluv.oauth.spotify.dto;
+
+public record ThumbNailResponse(
+    Integer height,
+    Integer width,
+    String url
+) {
+
+}
